fix(folder): make sure folders are handled correctly regadless of order
With the new priority sorting (done by getFolderCount() in RunboxWebmailAPI)
there is a risk that subfolders appear before parent folders in the
resulting list. The flattening algorithm in FolderListComponent doesn't
handle that correctly, breaks and displays nothing. This fixes it and
adds a testcase.

It feels like now that the folders are pre sorted the BFS loop (or what
looks like one :)) below could be somewhat simplified too – but that's
a task for another time.