fix(payments): Make the shopping cart more resilient
This includes a few things:

1) abandon /cart_POST API in favour of the improved
   /available_GET?pids=... endpoint.

This means we're using the same endpoint for loading the list
of available products and purchasing them later, which should minimize
the risk of something being addable to the cart and then not showing up.

2) Try to load products from paymentservice without calling the API.

Similar to the above: in the common case, the user will load click
something in another section to add it to the cart and then go to
the cart. We will now reuse the same product we showed them before,
meaning that we don't have to go to the API to load it again – which
should, again, minimize the chance that it fails to load and the cart
looks useless.

3) Show products even if they failed to load.

They'll show up as "Product #1234" and without a price: but the
determined user will still be able to buy them, and the transaction
should go through normally. This will not look pretty, but it will
actullly work if you go through with it.

Combined with the loading screen changes from the previous commits,
which will make a distinguishable difference between "there's nothing in
the cart" and "loading the cart failed", this will hopefully immunize
the cart to some of the bugs users have been experiencing.