refactor(authguard): Add a fast pass for previously logged in users
We used to hold each navigation request until we're sure that the user is
still logged in -- effectively making the user wait until /rest/v1/me
responds, adding hundreds of miliseconds (at best!) to each route change,
and making the app appear sluggish on operations as trivial as viewing
the drafts or going to /compose.

This implements a "fast pass" for the auth checker. If we previously saw
the user being logged in, we assume they still are and immediately
direct them to where they want to be. Meanwhile, under the hood, we
check if they're actually logged in -- and if not, we kick them out to
login like we used to.

For users not previously logged in, the navigation will block like it
used to. If it does not, the redirect from login to / would result in a
dysfunctional webmail, as nothing in the system would know who's coming.