feat(update-notifier): Include commit hash in app update data
This way we can display the upcoming version properly in the update
notes (as opposed to the old build timestamp that we used to show).

Also, knowing both the previous and the next commit hash (after this has
been deployed for a while) we'll be able to link users to changelogs
generated on the fly from a given commit range.