fix(canvastable): Make sure column widths are remembered and respected
This is mostly a refactor under the hood: what used to happen here is
the column widths would be set by the providers of the columns
themselves: xapian, rmmapi, or websocket. Separation of concerns (and
basic code hygiene :)) dictates that data providers should have nothing
to do with the interface layout, so this removes that capability of
theirs -- it's now all set in the canvastable itself. An added bonus
is the fact they're now actually consistent: previously, because of the
"decentralization" of their widths, every provider would set them to
slightly different values :)

The original reason (I assume...) for having them there was that the
data providers actually know which column is which: "a 4th column"
can mean different things depending on what we're displaying. To fix
this, column widths are now stored based on their name rather than their
index. This has a weakness -- the columns that have an empty name
(checkbox, flag, reply marker etc) are now treated as if they were the
same thing. This could be a problem if a user chose to resize only one
of them, but since they all only display a tiny icon, there's little
reasons to have them in different sizes. If we really wanted to support
that, a proper implementation would have to have unique IDs associated
with each of the columns, and based on this ID we would assign them
names and widths. Perhaps that's something worth doing in the future if
it turns out to be needed -- and if we don't replace this with angular's
virtual tables by then :)

Because the widths are now stored by name and not by index, the
localStorage setting for them has changed: which likely won't impact
anyone given that it used to be mostly broken anyway.