fix(calendar): Load timezones data from Icals instead of moment.js
This fixes a (real) problem of non-standard timezone names in events,
as well as a theoretical problem of moment.tz timezone data actually
differing from the one we encounter in real world.

It also bumps the cache version for calendars since we now need to store
the entire jcal (not just vevent) to preserve the timezone information
in local cache.

Tests fixes required by this also mean that we're now cheating in them
much less :)

Fixes GH-478.