fix(account): Fix an edgecase where payment would not start due to currency mismatch
From now on we'll rely on the `tx` object alone in all payment flows,
which should finally respect the currency that may have been forced into
USD (or something else) for a particular payment because of the products
being ordered.

We're also more strict now when loading the payment products, which
should fail earlier and produce better error messages.