feat(folderlist): Calculate folder counts from xapian if available
This approach is much faster and more consistent than before, but it
still suffers from a few flaws:

1) Index updates slightly later than the message flag is changed: so
right after the flag is changed, the folder counts are recalculated
according to the old index state, and only then the index is refreshed,
folder counts recalculated and the view refreshed.

2) Each index update calculates message counts for its own reasons,
and then it triggers a refreshFolderCounts() which calls SearchService's
getMessageCountsForFolder() multiple times -- in the end, we calculate
message counts multiple times for the same index state, which is
wasteful and unnecessary.

Ideally, the numbers would be updated locally each time the "seen" flag
is flipped, without waiting for the backend indexer to update us with
the new index. They would still be updated along with regular index
refreshes.