fix(canvastable): Be a bit more lenient when detecting message selection
Previously it'd require you to not do *any* mouse movements between
mousedown and mouseup (while selecting a single row), which would make
it appear inconsistent. This fixes is so that the selection works even
with a slight swipe, as long as the row and column is actually the same
before and after the drag.