fix(mailview): Remove deleted/moved messages before mesg list redraw
Fixes #713 (hopefully!)

The logic for delete/move messages was not removing the messages from
the separate "message display" list of ids that the canvas uses to
paint its list from. A regular repaint was trying to redraw the list
using the old list of ids, fetching data from the index which had
already been deleted, this errored out, quitting in the middle of
redrawing the message list.

This fix:
1) ensures we remove the items from the message display first
2) ensures the repaint continues even if it cant find some of the
message data.