feature(mailviewer): Store "show html images" per sender
The recent new feature to only show external images (et al) when
requested, implemented to speed up HTML display, is annoying some users.

This improvement allows users to save the "show external html images"
per sender (stored as a separate contact type in the dav contacts).

A new contact is only created if they do not already exist.

If/when the user later creates a contact with the same email as the
hidden type, we re-use the same contact and change it to the
"individual" type so that it will be displayed in Contacts.

Fixes #1249