fix(tooltips): Move context tool button tooltips to the left side
Currently tooltips show at the bottom of the button, blocking the button
bellow it.

Because of some weird behaviour, the tooltip does not vanish
when you move off it. This is probably due to an outdated material
version, because when reviewing the examples I do not experience
the same behaviour
https://v6.material.angular.io/components/tooltip/overview

As a temporary fix, I have moved the tooltips to the left.

See https://github.com/runbox/runbox7/issues/1439